requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
autopep8==2.1.0
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.8
Django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
filelock==3.13.3
identify==2.5.35
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pytest==8.0.2
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
psycopg2==2.9.9
django-cors-headers==4.3.1